refactoring: Split Inworld packet to separate class parts
