+++ packages/web-core/__tests__/entities/emotion_strength.entity.spec.ts
@@ -2,7 +2,7 @@ import { EmotionEventStrength } from '../../proto/ai/inworld/packets/packets.pb'
 import {
   EmotionStrength,
   EmotionStrengthCode,
-} from '../../src/entities/emotion_strength.entity';
+} from '../../src/entities/packets/emotion/emotion_strength.entity';
 
 const mappingTestTable = [
   {
