+++ packages/web-core/src/entities/continuation/previous_dialog.entity.ts
@@ -52,7 +52,7 @@ export class PreviousDialog {
         ({
           actor: DialogActor.toProto(item.talker),
           text: item.phrase,
-        } as DialogHistoryHistoryItem),
+        }) as DialogHistoryHistoryItem,
     );
 
     return { history };
