+++ packages/web-core/src/entities/packets/emotion/emotion_behavior.entity.ts
@@ -1,4 +1,4 @@
-import { EmotionEventSpaffCode } from '../../proto/ai/inworld/packets/packets.pb';
+import { EmotionEventSpaffCode } from '../../../../proto/ai/inworld/packets/packets.pb';
 
 export enum EmotionBehaviorCode {
   NEUTRAL = 'NEUTRAL',
