+++ packages/web-core/src/index.ts
@@ -39,26 +39,27 @@ import {
   SessionContinuationProps,
 } from './entities/continuation/session_continuation.entity';
 import {
+  AdditionalPhonemeInfo,
+  AudioEvent,
+} from './entities/packets/audio.entity';
+import { CancelResponsesEvent } from './entities/packets/cancel_responses.entity';
+import { EmotionEvent } from './entities/packets/emotion/emotion.entity';
+import {
   EmotionBehavior,
   EmotionBehaviorCode,
-} from './entities/emotion_behavior.entity';
+} from './entities/packets/emotion/emotion_behavior.entity';
 import {
   EmotionStrength,
   EmotionStrengthCode,
-} from './entities/emotion_strength.entity';
+} from './entities/packets/emotion/emotion_strength.entity';
+import { InworldPacket } from './entities/packets/inworld_packet.entity';
+import { PacketId } from './entities/packets/packet_id.entity';
+import { Actor, Routing } from './entities/packets/routing.entity';
+import { TextEvent } from './entities/packets/text.entity';
 import {
-  Actor,
-  AdditionalPhonemeInfo,
-  AudioEvent,
-  CancelResponsesEvent,
-  EmotionEvent,
-  InworldPacket,
-  PacketId,
-  Routing,
-  TextEvent,
   TriggerEvent,
   TriggerParameter,
-} from './entities/inworld_packet.entity';
+} from './entities/packets/trigger.entity';
 import { SessionToken } from './entities/session_token.entity';
 import { InworldConnectionService } from './services/inworld_connection.service';
 import { SessionState } from './services/pb/state_serialization.service';
